Ensure GetBestRect is larger or equal to the ideal rect
[NSView frame] returns floating-point pixel values for the frame
size.  wxWidgets only deals with integer pixel values, but this
can cause issues if the truncation results in a smaller rect
than the ideal.  For widgets like wxStaticText, setting the
size even a tiny bit smaller than the rect returned by
[NSView frame] can result in text truncation.

This problem is new in 10.15.  I haven't investigated whether
the return value of [NSView frame] changed in 10.15, or whether
the change is elsewhere (such as in the way that static text
widgets truncate text if the size is too small, etc).