Add a function to save a bitmap callable from debugger
This allows to dump a bitmap to a disk file in order to examine it
outside of the debugger when debugging.

Note that at least MSVC and gdb are not smart enough to allow calling
wxBitmap::SaveFile() directly.

(cherry picked from commit 2d6817b0a8799f428d64207ab1d8bf030a8b4e36)