Fix using danging pointer in wxRendererMac::DrawMacHeader()
Keeping a pointer to NSString corresponding to temporary wxCFStringRef
was wrong, as this pointer became dangling as soon as the temporary
object was destroyed.

Fix this by assigning this pointer to NSTableHeaderCell.cell, which
increments its reference count and so keeps it alive for long enough.

This incidentally also fixes wrongly using left alignment when right or
center alignment was specified, as the value assigned to cell.alignment
was always overwritten with NSTextAlignment after setting it.

Closes #24511.

Co-authored-by: Marek Roszko <mark.roszko@gmail.com>
(cherry picked from commit 0c07050569c61ebf8f217ae72b7cee05296cfee1)